Fix newline parsing
Standalone LF character parsing was taking precedence over LFCR.
LFCR should be attmpted first, then on failure, LF alone can match.

Also separates tests between StringParsing and core parser combinator
functions.
